organize
organized pivot code
