+++ src/main/java/frc/robot/RobotContainer.java
@@ -20,9 +20,9 @@ import frc.robot.Constants.ManipulatorConstants;
 import frc.robot.Constants.OIConstants;
 import frc.robot.commands.setXCommand;
 import frc.robot.commands.manipulator.Intake;
-import frc.robot.commands.manipulator.PivotDown;
-import frc.robot.commands.manipulator.PivotUp;
 import frc.robot.commands.manipulator.Shooter;
+import frc.robot.commands.manipulator.pivot.PivotDown;
+import frc.robot.commands.manipulator.pivot.PivotUp;
 import frc.robot.commands.vision.AlignScore;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
